refactor(knockdog): geo 관련 공통 유틸 함수 추가
