+++ apps/knockdog/src/shared/lib/geolocation/index.ts
@@ -1,5 +1,5 @@
-export * from './is';
-
+export { isValidCoord, isEqualCoord } from './is';
+export { serializeCoords } from './serialize';
 export { getCurrentLocation } from './getCurrentLocation';
 export { useCurrentLocation } from './useCurrentLocation';
 export { useBasePoint } from './useBasePoint';
